Add GitHub funding information
I’m part of the GitHub sponsors program, so people are able to fund
development of packages that I own that they find useful.  While this
is a small and niche package, there is no issue with including the
funding information.  This patch adds a `FUNDING.yml` file in the
format required by GitHub to include this information on the
repository overview.

Signed-off-by: Patrick M. Niedzielski <patrick@pniedzielski.net>
License: BSD-3-Clause